lunes entretenido

+++ libft/ft_bzero.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
 	size_t	count;
